Provides fragment ends analysis

Calculates the number of fragment ends and the Windowed Protection Score (WPS) in genomic tiles within targets

analyze_fragmentation(bam, targets, tag = "", window_size = 120,
  step_size = 5, min_size = 120, max_size = 180, ...)

Arguments

bam
the input bam file

targets
The targets to restrict the windows within. Must have the columns chr, start and end. In case of whole-genome, specify full chromosomes targets.

tag
the RG tag if the bam has more than one sample.

window_size
The window (bin) size to use within the targets

step_size
The step size to use in case of overlapping bins.

min_size
Restrict fragments to this minimum size.

max_size
Restrict fragments to this maximum size.

...
Other parameters passed to get_fragment_size

Value

a data frame with the first three columns having the bins coordinates and other columns having the WPS (raw and adjusted by coverage) and number of fragment ends (raw and adjusted by coverage).

Details

Fragment length will extracted from the bam file according to the parameters passed to get_fragment_size, and the number of fragment ends, and the Windowed Protection Score (WPS) will be computed in the binned input targets. Binning is done according to the window_size and step_size parameters.

WPS is defined as the number of fragments completely spanning a window (bin) minus the number of fragments with an endpoint within the same window as reported by Snyder et al., Cell 2016.

The output include both the fragment end counts and the WPS in their raw format as well as after adjustment by coverage in the bin.

Minimum and maximum bounds of the fragment size are applied before computing WPS and fragment ends counts.

See also

Examples

# \donttest{
data("targets", package = "ctDNAtools")
bamN1 <- system.file("extdata", "N1.bam", package = "ctDNAtools")

## basic usage
analyze_fragmentation(bam = bamN1, targets = targets)

#>       chr     start       end  WPS WPS_adjusted n_fragment_ends
#> 1   chr14 106327303 106327423 -155   -1.0000000             155
#> 2   chr14 106327308 106327428 -165   -1.0000000             165
#> 3   chr14 106327313 106327433 -170   -1.0000000             170
#> 4   chr14 106327318 106327438 -176   -1.0000000             176
#> 5   chr14 106327323 106327443 -171   -0.9661017             174
#> 6   chr14 106327328 106327448 -172   -0.9347826             179
#> 7   chr14 106327333 106327453 -170   -0.9239130             177
#> 8   chr14 106327338 106327458 -157   -0.8395722             172
#> 9   chr14 106327343 106327463 -155   -0.8288770             171
#> 10  chr14 106327348 106327468 -145   -0.7591623             172
#> 11  chr14 106327353 106327473 -139   -0.7202073             168
#> 12  chr14 106327358 106327478 -137   -0.7025641             166
#> 13  chr14 106327363 106327483 -134   -0.6836735             171
#> 14  chr14 106327368 106327488 -133   -0.6751269             165
#> 15  chr14 106327373 106327493 -119   -0.5804878             164
#> 16  chr14 106327378 106327498 -119   -0.5693780             167
#> 17  chr14 106327383 106327503 -127   -0.6018957             169
#> 18  chr14 106327388 106327508  -65   -0.3080569             138
#> 19  chr14 106327393 106327513  -47   -0.2165899             135
#> 20  chr14 106327398 106327518  -35   -0.1583710             133
#> 21  chr14 106327403 106327523  -41   -0.1744681             138
#> 22  chr14 106327408 106327528  -62   -0.2605042             151
#> 23  chr14 106327413 106327533  -64   -0.2644628             165
#> 24  chr14 106327418 106327538  -85   -0.3281853             175
#> 25  chr14 106327423 106327543 -109   -0.4113208             187
#> 26  chr14 106327428 106327548 -140   -0.5263158             206
#> 27  chr14 106327433 106327553 -157   -0.5709091             217
#> 28  chr14 106327438 106327558 -178   -0.6312057             233
#> 29  chr14 106327443 106327563 -190   -0.6643357             239
#> 30  chr14 106327448 106327568 -213   -0.7123746             256
#> 31  chr14 106327453 106327573 -227   -0.7541528             264
#> 32  chr14 106327458 106327578 -228   -0.7402597             270
#> 33  chr14 106327463 106327583 -250   -0.7812500             287
#> 34  chr14 106327468 106327588 -270   -0.8035714             304
#> 35  chr14 106327473 106327593 -304   -0.8735632             328
#> 36  chr14 106327478 106327598 -312   -0.8813559             333
#> 37  chr14 106327483 106327603 -320   -0.8888889             343
#> 38  chr14 106327488 106327608 -337   -0.9335180             350
#> 39  chr14 106327493 106327613 -323   -0.8947368             343
#> 40  chr14 106327498 106327618 -325   -0.8904110             349
#> 41  chr14 106327503 106327623 -321   -0.8891967             341
#> 42  chr14 106327508 106327628 -327   -0.9008264             345
#> 43  chr14 106327513 106327633 -323   -0.8753388             346
#> 44  chr14 106327518 106327638 -316   -0.8586957             343
#> 45  chr14 106327523 106327643 -292   -0.8021978             334
#> 46  chr14 106327528 106327648 -279   -0.7859155             317
#> 47  chr14 106327533 106327653 -268   -0.7613636             310
#> 48  chr14 106327538 106327658 -262   -0.7616279             310
#> 49  chr14 106327543 106327663 -235   -0.7142857             285
#> 50  chr14 106327548 106327668 -217   -0.7114754             261
#> 51  chr14 106327553 106327673 -194   -0.6466667             249
#> 52  chr14 106327558 106327678 -173   -0.5945017             233
#> 53  chr14 106327563 106327683 -166   -0.5763889             228
#> 54  chr14 106327568 106327688 -136   -0.4892086             207
#> 55  chr14 106327573 106327693 -144   -0.5217391             216
#> 56  chr14 106327578 106327698 -160   -0.5755396             222
#> 57  chr14 106327583 106327703 -154   -0.5422535             223
#> 58  chr14 106327588 106327708 -154   -0.5422535             226
#> 59  chr14 106327593 106327713 -121   -0.4464945             199
#> 60  chr14 106327598 106327718 -114   -0.4253731             192
#> 61  chr14 106327603 106327723 -103   -0.3857678             188
#> 62  chr14 106327608 106327728 -117   -0.4448669             191
#> 63  chr14 106327613 106327733 -145   -0.5272727             211
#> 64  chr14 106327618 106327738 -140   -0.5109489             207
#> 65  chr14 106327623 106327743 -137   -0.4981818             207
#> 66  chr14 106327628 106327748 -143   -0.5238095             209
#> 67  chr14 106327633 106327753 -147   -0.5384615             214
#> 68  chr14 106327638 106327758 -190   -0.6834532             235
#> 69  chr14 106327643 106327763 -193   -0.6868327             238
#> 70  chr14 106327648 106327768 -222   -0.7655172             258
#> 71  chr14 106327653 106327773 -234   -0.7905405             265
#> 72  chr14 106327658 106327778 -231   -0.7938144             262
#> 73  chr14 106327663 106327783 -246   -0.8092105             276
#> 74  chr14 106327668 106327788 -251   -0.8283828             278
#> 75  chr14 106327673 106327793 -256   -0.8152866             285
#> 76  chr14 106327678 106327798 -263   -0.8349206             289
#> 77  chr14 106327683 106327803 -259   -0.8222222             287
#> 78  chr14 106327688 106327808 -268   -0.8589744             290
#> 79  chr14 106327693 106327813 -258   -0.8376623             284
#> 80  chr14 106327698 106327818 -246   -0.8310811             271
#> 81  chr14 106327703 106327823 -235   -0.8020478             268
#> 82  chr14 106327708 106327828 -224   -0.7777778             256
#> 83  chr14 106327713 106327833 -228   -0.7916667             258
#> 84  chr14 106327718 106327838 -227   -0.7964912             257
#> 85  chr14 106327723 106327843 -234   -0.8239437             259
#> 86  chr14 106327728 106327848 -225   -0.8064516             256
#> 87  chr14 106327733 106327853 -198   -0.7333333             235
#> 88  chr14 106327738 106327858 -202   -0.7481481             239
#> 89  chr14 106327743 106327863 -204   -0.7555556             237
#> 90  chr14 106327748 106327868 -198   -0.7500000             234
#> 91  chr14 106327753 106327873 -195   -0.7647059             225
#> 92  chr14 106327758 106327878 -179   -0.7489540             209
#> 93  chr14 106327763 106327883 -168   -0.7118644             202
#> 94  chr14 106327768 106327888 -140   -0.6140351             188
#> 95  chr14 106327773 106327893 -132   -0.5789474             180
#> 96  chr14 106327778 106327898 -109   -0.4844444             173
#> 97  chr14 106327783 106327903  -93   -0.4096916             166
#> 98  chr14 106327788 106327908 -106   -0.4732143             166
#> 99  chr14 106327793 106327913  -87   -0.3866667             156
#> 100 chr14 106327798 106327918  -90   -0.4054054             156
#> 101 chr14 106327803 106327923  -88   -0.3963964             156
#> 102 chr14 106327808 106327928  -96   -0.4444444             157
#> 103 chr14 106327813 106327933 -122   -0.5495495             174
#> 104 chr14 106327818 106327938 -136   -0.6181818             187
#> 105 chr14 106327823 106327943 -170   -0.7657658             197
#> 106 chr14 106327828 106327948 -160   -0.7017544             198
#> 107 chr14 106327833 106327953 -172   -0.7413793             202
#> 108 chr14 106327838 106327958 -186   -0.7622951             217
#> 109 chr14 106327843 106327963 -197   -0.7975709             224
#> 110 chr14 106327848 106327968 -187   -0.7391304             224
#> 111 chr14 106327853 106327973 -190   -0.7600000             220
#> 112 chr14 106327858 106327978 -180   -0.7200000             220
#> 113 chr14 106327863 106327983 -181   -0.7269076             216
#> 114 chr14 106327868 106327988 -187   -0.7510040             218
#> 115 chr14 106327873 106327993 -206   -0.8373984             228
#> 116 chr14 106327878 106327998 -205   -0.8367347             226
#> 117 chr14 106327883 106328003 -204   -0.8429752             223
#> 118 chr14 106327888 106328008 -200   -0.8264463             228
#> 119 chr14 106327893 106328013 -196   -0.8166667             218
#> 120 chr14 106327898 106328018 -193   -0.8143460             216
#> 121 chr14 106327903 106328023 -185   -0.7939914             214
#> 122 chr14 106327908 106328028 -188   -0.8468468             206
#> 123 chr14 106327913 106328033 -185   -0.8371041             205
#> 124 chr14 106327918 106328038 -176   -0.8148148             199
#> 125 chr14 106327923 106328043 -181   -0.8418605             198
#> 126 chr14 106327928 106328048 -170   -0.8173077             190
#> 127 chr14 106327933 106328053 -145   -0.7435897             173
#> 128 chr14 106327938 106328058 -143   -0.7729730             164
#> 129 chr14 106327943 106328063 -108   -0.6506024             139
#> 130 chr14 106327948 106328068  -95   -0.5828221             133
#> 131 chr14 106327953 106328073  -79   -0.5031847             123
#> 132 chr14 106327958 106328078  -56   -0.3636364             107
#> 133 chr14 106327963 106328083  -36   -0.2465753              91
#> 134 chr14 106327968 106328088  -22   -0.1506849              93
#> 135 chr14 106327973 106328093  -25   -0.1748252              84
#> 136 chr14 106327978 106328098  -24   -0.1690141              83
#> 137 chr14 106327983 106328103  -20   -0.1428571              80
#> 138 chr14 106327988 106328108  -48   -0.3529412              95
#> 139 chr14 106327993 106328113  -43   -0.3440000              86
#> 140 chr14 106327998 106328118  -58   -0.4833333              89
#> 141 chr14 106328003 106328123  -77   -0.6470588              98
#> 142 chr14 106328008 106328128  -85   -0.7264957             101
#> 143 chr14 106328013 106328133  -84   -0.7636364              97
#> 144 chr14 106328018 106328138  -87   -0.7981651              99
#> 145 chr14 106328023 106328143 -100   -0.9259259             106
#> 146 chr14 106328028 106328148 -100   -1.0000000             100
#> 147 chr14 106328033 106328153  -99   -1.0000000              99
#> 148 chr14 106328038 106328158  -97   -1.0000000              97
#> 149 chr14 106328043 106328163  -94   -1.0000000              94
#> 150 chr14 106328048 106328168  -94   -1.0000000              94
#> 151 chr14 106328053 106328173  -93   -1.0000000              93
#> 152 chr14 106328058 106328178  -89   -1.0000000              89
#> 153 chr14 106328063 106328183  -88   -1.0000000              88
#> 154 chr14 106328068 106328188  -84   -1.0000000              84
#> 155 chr14 106328073 106328193  -83   -1.0000000              83
#> 156 chr14 106328078 106328198  -80   -1.0000000              80
#> 157 chr14 106328083 106328203  -78   -1.0000000              78
#> 158 chr14 106328088 106328208  -76   -1.0000000              76
#> 159 chr14 106328093 106328213  -72   -1.0000000              72
#> 160 chr14 106328098 106328218  -68   -1.0000000              68
#> 161 chr14 106328103 106328223  -66   -1.0000000              66
#> 162 chr14 106328108 106328228  -50   -1.0000000              50
#> 163 chr14 106328113 106328233  -46   -1.0000000              46
#> 164 chr14 106328118 106328238  -31   -1.0000000              31
#> 165 chr14 106328123 106328243  -21   -1.0000000              21
#> 166 chr14 106328128 106328248  -16   -1.0000000              16
#> 167 chr14 106328133 106328253  -13   -1.0000000              13
#> 168 chr14 106328138 106328258  -11   -1.0000000              11
#> 169 chr14 106328143 106328263   -4   -1.0000000               4
#> 170 chr14 106328148 106328268    0    0.0000000               0
#> 171 chr14 106328153 106328273    0    0.0000000               0
#> 172 chr14 106328158 106328278    0    0.0000000               0
#> 173 chr14 106328163 106328283    0    0.0000000               0
#> 174 chr14 106328168 106328288    0    0.0000000               0
#>     n_fragment_ends_adjusted n_reads
#> 1                  1.0000000     155
#> 2                  1.0000000     165
#> 3                  1.0000000     170
#> 4                  1.0000000     176
#> 5                  0.9830508     177
#> 6                  0.9728261     184
#> 7                  0.9619565     184
#> 8                  0.9197861     187
#> 9                  0.9144385     187
#> 10                 0.9005236     191
#> 11                 0.8704663     193
#> 12                 0.8512821     195
#> 13                 0.8724490     196
#> 14                 0.8375635     197
#> 15                 0.8000000     205
#> 16                 0.7990431     209
#> 17                 0.8009479     211
#> 18                 0.6540284     211
#> 19                 0.6221198     217
#> 20                 0.6018100     221
#> 21                 0.5872340     235
#> 22                 0.6344538     238
#> 23                 0.6818182     242
#> 24                 0.6756757     259
#> 25                 0.7056604     265
#> 26                 0.7744361     266
#> 27                 0.7890909     275
#> 28                 0.8262411     282
#> 29                 0.8356643     286
#> 30                 0.8561873     299
#> 31                 0.8770764     301
#> 32                 0.8766234     308
#> 33                 0.8968750     320
#> 34                 0.9047619     336
#> 35                 0.9425287     348
#> 36                 0.9406780     354
#> 37                 0.9527778     360
#> 38                 0.9695291     361
#> 39                 0.9501385     361
#> 40                 0.9561644     365
#> 41                 0.9445983     361
#> 42                 0.9504132     363
#> 43                 0.9376694     369
#> 44                 0.9320652     368
#> 45                 0.9175824     364
#> 46                 0.8929577     355
#> 47                 0.8806818     352
#> 48                 0.9011628     344
#> 49                 0.8662614     329
#> 50                 0.8557377     305
#> 51                 0.8300000     300
#> 52                 0.8006873     291
#> 53                 0.7916667     288
#> 54                 0.7446043     278
#> 55                 0.7826087     276
#> 56                 0.7985612     278
#> 57                 0.7852113     284
#> 58                 0.7957746     284
#> 59                 0.7343173     271
#> 60                 0.7164179     268
#> 61                 0.7041199     267
#> 62                 0.7262357     263
#> 63                 0.7672727     275
#> 64                 0.7554745     274
#> 65                 0.7527273     275
#> 66                 0.7655678     273
#> 67                 0.7838828     273
#> 68                 0.8453237     278
#> 69                 0.8469751     281
#> 70                 0.8896552     290
#> 71                 0.8952703     296
#> 72                 0.9003436     291
#> 73                 0.9078947     304
#> 74                 0.9174917     303
#> 75                 0.9076433     314
#> 76                 0.9174603     315
#> 77                 0.9111111     315
#> 78                 0.9294872     312
#> 79                 0.9220779     308
#> 80                 0.9155405     296
#> 81                 0.9146758     293
#> 82                 0.8888889     288
#> 83                 0.8958333     288
#> 84                 0.9017544     285
#> 85                 0.9119718     284
#> 86                 0.9175627     279
#> 87                 0.8703704     270
#> 88                 0.8851852     270
#> 89                 0.8777778     270
#> 90                 0.8863636     264
#> 91                 0.8823529     255
#> 92                 0.8744770     239
#> 93                 0.8559322     236
#> 94                 0.8245614     228
#> 95                 0.7894737     228
#> 96                 0.7688889     225
#> 97                 0.7312775     227
#> 98                 0.7410714     224
#> 99                 0.6933333     225
#> 100                0.7027027     222
#> 101                0.7027027     222
#> 102                0.7268519     216
#> 103                0.7837838     222
#> 104                0.8500000     220
#> 105                0.8873874     222
#> 106                0.8684211     228
#> 107                0.8706897     232
#> 108                0.8893443     244
#> 109                0.9068826     247
#> 110                0.8853755     253
#> 111                0.8800000     250
#> 112                0.8800000     250
#> 113                0.8674699     249
#> 114                0.8755020     249
#> 115                0.9268293     246
#> 116                0.9224490     245
#> 117                0.9214876     242
#> 118                0.9421488     242
#> 119                0.9083333     240
#> 120                0.9113924     237
#> 121                0.9184549     233
#> 122                0.9279279     222
#> 123                0.9276018     221
#> 124                0.9212963     216
#> 125                0.9209302     215
#> 126                0.9134615     208
#> 127                0.8871795     195
#> 128                0.8864865     185
#> 129                0.8373494     166
#> 130                0.8159509     163
#> 131                0.7834395     157
#> 132                0.6948052     154
#> 133                0.6232877     146
#> 134                0.6369863     146
#> 135                0.5874126     143
#> 136                0.5845070     142
#> 137                0.5714286     140
#> 138                0.6985294     136
#> 139                0.6880000     125
#> 140                0.7416667     120
#> 141                0.8235294     119
#> 142                0.8632479     117
#> 143                0.8818182     110
#> 144                0.9082569     109
#> 145                0.9814815     108
#> 146                1.0000000     100
#> 147                1.0000000      99
#> 148                1.0000000      97
#> 149                1.0000000      94
#> 150                1.0000000      94
#> 151                1.0000000      93
#> 152                1.0000000      89
#> 153                1.0000000      88
#> 154                1.0000000      84
#> 155                1.0000000      83
#> 156                1.0000000      80
#> 157                1.0000000      78
#> 158                1.0000000      76
#> 159                1.0000000      72
#> 160                1.0000000      68
#> 161                1.0000000      66
#> 162                1.0000000      50
#> 163                1.0000000      46
#> 164                1.0000000      31
#> 165                1.0000000      21
#> 166                1.0000000      16
#> 167                1.0000000      13
#> 168                1.0000000      11
#> 169                1.0000000       4
#> 170                0.0000000       0
#> 171                0.0000000       0
#> 172                0.0000000       0
#> 173                0.0000000       0
#> 174                0.0000000       0

## more options
analyze_fragmentation(
  bam = bamN1, targets = targets,
  step_size = 10, window_size = 50
)

#>      chr     start       end WPS WPS_adjusted n_fragment_ends
#> 1  chr14 106327373 106327423 -51  -0.32903226             104
#> 2  chr14 106327383 106327433 -46  -0.27058824             108
#> 3  chr14 106327393 106327443  37   0.20903955              71
#> 4  chr14 106327403 106327453  74   0.40217391              55
#> 5  chr14 106327413 106327463  93   0.49732620              54
#> 6  chr14 106327423 106327473 117   0.60621762              38
#> 7  chr14 106327433 106327483 134   0.68367347              33
#> 8  chr14 106327443 106327493 131   0.63902439              38
#> 9  chr14 106327453 106327503 117   0.55450237              47
#> 10 chr14 106327463 106327513 113   0.52073733              54
#> 11 chr14 106327473 106327523  87   0.37021277              76
#> 12 chr14 106327483 106327533  53   0.22362869              97
#> 13 chr14 106327493 106327543   0   0.00000000             129
#> 14 chr14 106327503 106327553 -55  -0.21568627             156
#> 15 chr14 106327513 106327563 -88  -0.33333333             177
#> 16 chr14 106327523 106327573 -93  -0.34572491             187
#> 17 chr14 106327533 106327583 -89  -0.33085502             181
#> 18 chr14 106327543 106327593 -83  -0.30627306             179
#> 19 chr14 106327553 106327603 -56  -0.22047244             160
#> 20 chr14 106327563 106327613 -41  -0.17012448             141
#> 21 chr14 106327573 106327623 -14  -0.05982906             125
#> 22 chr14 106327583 106327633  13   0.05439331             114
#> 23 chr14 106327593 106327643  73   0.32444444              78
#> 24 chr14 106327603 106327653 103   0.45374449              62
#> 25 chr14 106327613 106327663  84   0.38181818              69
#> 26 chr14 106327623 106327673  98   0.44545455              61
#> 27 chr14 106327633 106327683 109   0.48878924              58
#> 28 chr14 106327643 106327693  98   0.43750000              68
#> 29 chr14 106327653 106327703  60   0.25423729              91
#> 30 chr14 106327663 106327713  42   0.18750000              92
#> 31 chr14 106327673 106327723  35   0.15837104              96
#> 32 chr14 106327683 106327733 -12  -0.05084746             125
#> 33 chr14 106327693 106327743  -9  -0.03896104             122
#> 34 chr14 106327703 106327753   4   0.01886792             112
#> 35 chr14 106327713 106327763 -27  -0.13043478             118
#> 36 chr14 106327723 106327773 -71  -0.33333333             142
#> 37 chr14 106327733 106327783 -49  -0.22580645             135
#> 38 chr14 106327743 106327793 -56  -0.25000000             140
#> 39 chr14 106327753 106327803 -56  -0.26415094             134
#> 40 chr14 106327763 106327813 -27  -0.14136126             109
#> 41 chr14 106327773 106327823   2   0.01136364              87
#> 42 chr14 106327783 106327833  36   0.19565217              76
#> 43 chr14 106327793 106327843  51   0.27567568              67
#> 44 chr14 106327803 106327853  44   0.23655914              72
#> 45 chr14 106327813 106327863  36   0.19565217              74
#> 46 chr14 106327823 106327873  46   0.26136364              65
#> 47 chr14 106327833 106327883  54   0.30681818              61
#> 48 chr14 106327843 106327893  56   0.30769231              65
#> 49 chr14 106327853 106327903  61   0.34078212              63
#> 50 chr14 106327863 106327913  56   0.32183908              59
#> 51 chr14 106327873 106327923  46   0.26436782              64
#> 52 chr14 106327883 106327933  16   0.09195402              81
#> 53 chr14 106327893 106327943 -25  -0.13812155             104
#> 54 chr14 106327903 106327953 -43  -0.22751323             118
#> 55 chr14 106327913 106327963 -68  -0.34343434             135
#> 56 chr14 106327923 106327973 -70  -0.34653465             136
#> 57 chr14 106327933 106327983 -45  -0.23809524             120
#> 58 chr14 106327943 106327993 -29  -0.18012422              99
#> 59 chr14 106327953 106328003  -7  -0.04458599              85
#> 60 chr14 106327963 106328013  28   0.19178082              59
#> 61 chr14 106327973 106328023  47   0.32867133              51
#> 62 chr14 106327983 106328033  46   0.32857143              47
#> 63 chr14 106327993 106328043  53   0.42400000              37
#> 64 chr14 106328003 106328053  67   0.56302521              27
#> 65 chr14 106328013 106328063  66   0.60000000              22
#> 66 chr14 106328023 106328073  58   0.53703704              27
#> 67 chr14 106328033 106328083  57   0.57575758              21
#> 68 chr14 106328043 106328093  50   0.53191489              22
#> 69 chr14 106328053 106328103  39   0.41935484              27
#> 70 chr14 106328063 106328113   4   0.04545455              43
#> 71 chr14 106328073 106328123 -41  -0.49397590              62
#> 72 chr14 106328083 106328133 -52  -0.66666667              65
#> 73 chr14 106328093 106328143 -64  -0.88888889              70
#>    n_fragment_ends_adjusted n_reads
#> 1                 0.6709677     155
#> 2                 0.6352941     170
#> 3                 0.4011299     177
#> 4                 0.2989130     184
#> 5                 0.2887701     187
#> 6                 0.1968912     193
#> 7                 0.1683673     196
#> 8                 0.1853659     205
#> 9                 0.2227488     211
#> 10                0.2488479     217
#> 11                0.3234043     235
#> 12                0.4092827     237
#> 13                0.5039062     256
#> 14                0.6117647     255
#> 15                0.6704545     264
#> 16                0.6951673     269
#> 17                0.6728625     269
#> 18                0.6605166     271
#> 19                0.6299213     254
#> 20                0.5850622     241
#> 21                0.5341880     234
#> 22                0.4769874     239
#> 23                0.3466667     225
#> 24                0.2731278     227
#> 25                0.3136364     220
#> 26                0.2772727     220
#> 27                0.2600897     223
#> 28                0.3035714     224
#> 29                0.3855932     236
#> 30                0.4107143     224
#> 31                0.4343891     221
#> 32                0.5296610     236
#> 33                0.5281385     231
#> 34                0.5283019     212
#> 35                0.5700483     207
#> 36                0.6666667     213
#> 37                0.6221198     217
#> 38                0.6250000     224
#> 39                0.6320755     212
#> 40                0.5706806     191
#> 41                0.4943182     176
#> 42                0.4130435     184
#> 43                0.3621622     185
#> 44                0.3870968     186
#> 45                0.4021739     184
#> 46                0.3693182     176
#> 47                0.3465909     176
#> 48                0.3571429     182
#> 49                0.3519553     179
#> 50                0.3390805     174
#> 51                0.3678161     174
#> 52                0.4655172     174
#> 53                0.5745856     181
#> 54                0.6243386     189
#> 55                0.6818182     198
#> 56                0.6732673     202
#> 57                0.6349206     189
#> 58                0.6149068     161
#> 59                0.5414013     157
#> 60                0.4041096     146
#> 61                0.3566434     143
#> 62                0.3357143     140
#> 63                0.2960000     125
#> 64                0.2268908     119
#> 65                0.2000000     110
#> 66                0.2500000     108
#> 67                0.2121212      99
#> 68                0.2340426      94
#> 69                0.2903226      93
#> 70                0.4886364      88
#> 71                0.7469880      83
#> 72                0.8333333      78
#> 73                0.9722222      72

# }