Gets histogram of fragment lengths from a bam file

The function first extracts fragment length from a bam file then computes the histogram over defined bins. If normalized is TRUE, the counts per bin will be normalized to the total read counts. Optionally, it can computes the histogram of fragment lengths only for mutated reads (confirmed ctDNA molecules).

bin_fragment_size(bam, mutations = NULL, targets = NULL, tag = "",
  bin_size = 2, custom_bins = NULL, normalized = FALSE,
  min_size = 1, max_size = 400, ...)

Arguments

bam
path to bam file.

mutations
An optional data frame with mutations. Must have the columns CHROM, POS, REF, ALT.

targets
a data frame with the target regions to restrict the reads in the bam. Must have three columns: chr, start and end

tag
the RG tag if the bam has more than one sample.

bin_size
the width of the bin (breaks) of the histogram.

custom_bins
A numeric vector for custom breaks to bin the histogram of fragment length. Over-rides bin_size.

normalized
A logical, whether to normalize the counts to the total number of reads.

min_size
Integer with the lowest fragment length.

max_size
Integer with the highest fragment length.

...
Other parameters passed to get_fragment_size.

Value

A data frame with one column for the used breaks and one having the histogram (normalized) counts. If mutations is supplied, the output will have one breaks column and three columns corresponding to variant allele reads, reference allele reads, and other reads. Each row has the count of fragment lengths within the bin and optionally normalized by the total number of reads.

Details

Fragment length will extracted from the bam file according to the parameters passed to get_fragment_size, and histogram counts (optionally normalized to total counts) are computed. Both equal histogram bins via bin_size and manually customized bins via custom_bins are supported.

By using an input mutations, the function will bin separately the reads that support variant alleles, reference alleles and other reads.

See also

Examples

# \donttest{
data("targets", package = "ctDNAtools")
data("mutations", package = "ctDNAtools")
bamT1 <- system.file("extdata", "T1.bam", package = "ctDNAtools")

## basic usage
bin_fragment_size(bam = bamT1)

#>      Breaks T1
#> 1       1_3  0
#> 2       3_5  0
#> 3       5_7  0
#> 4       7_9  0
#> 5      9_11  0
#> 6     11_13  0
#> 7     13_15  0
#> 8     15_17  0
#> 9     17_19  0
#> 10    19_21  0
#> 11    21_23  0
#> 12    23_25  0
#> 13    25_27  0
#> 14    27_29  0
#> 15    29_31  0
#> 16    31_33  0
#> 17    33_35  0
#> 18    35_37  0
#> 19    37_39  0
#> 20    39_41  0
#> 21    41_43  0
#> 22    43_45  0
#> 23    45_47  0
#> 24    47_49  0
#> 25    49_51  0
#> 26    51_53  0
#> 27    53_55  0
#> 28    55_57  0
#> 29    57_59  0
#> 30    59_61  0
#> 31    61_63  0
#> 32    63_65  0
#> 33    65_67  0
#> 34    67_69  0
#> 35    69_71  0
#> 36    71_73  0
#> 37    73_75  0
#> 38    75_77  0
#> 39    77_79  0
#> 40    79_81  0
#> 41    81_83  0
#> 42    83_85  0
#> 43    85_87  0
#> 44    87_89  0
#> 45    89_91  0
#> 46    91_93  0
#> 47    93_95  0
#> 48    95_97  0
#> 49    97_99  0
#> 50   99_101  0
#> 51  101_103  0
#> 52  103_105  1
#> 53  105_107  0
#> 54  107_109  0
#> 55  109_111  0
#> 56  111_113  3
#> 57  113_115  0
#> 58  115_117  5
#> 59  117_119  1
#> 60  119_121  3
#> 61  121_123  9
#> 62  123_125  1
#> 63  125_127  1
#> 64  127_129  3
#> 65  129_131  5
#> 66  131_133  9
#> 67  133_135  4
#> 68  135_137  6
#> 69  137_139  7
#> 70  139_141  9
#> 71  141_143 13
#> 72  143_145 15
#> 73  145_147 13
#> 74  147_149 17
#> 75  149_151 28
#> 76  151_153 19
#> 77  153_155 21
#> 78  155_157 24
#> 79  157_159 34
#> 80  159_161 34
#> 81  161_163 30
#> 82  163_165 61
#> 83  165_167 52
#> 84  167_169 64
#> 85  169_171 44
#> 86  171_173 55
#> 87  173_175 39
#> 88  175_177 38
#> 89  177_179 45
#> 90  179_181 43
#> 91  181_183 34
#> 92  183_185 22
#> 93  185_187 27
#> 94  187_189 35
#> 95  189_191 21
#> 96  191_193 14
#> 97  193_195 21
#> 98  195_197 15
#> 99  197_199 16
#> 100 199_201  8
#> 101 201_203 11
#> 102 203_205  8
#> 103 205_207  4
#> 104 207_209  9
#> 105 209_211  4
#> 106 211_213 13
#> 107 213_215  7
#> 108 215_217  5
#> 109 217_219  7
#> 110 219_221  5
#> 111 221_223  4
#> 112 223_225  4
#> 113 225_227  5
#> 114 227_229  3
#> 115 229_231  1
#> 116 231_233  3
#> 117 233_235  4
#> 118 235_237  5
#> 119 237_239  1
#> 120 239_241  5
#> 121 241_243  1
#> 122 243_245  2
#> 123 245_247  3
#> 124 247_249  2
#> 125 249_251  1
#> 126 251_253  2
#> 127 253_255  0
#> 128 255_257  1
#> 129 257_259  2
#> 130 259_261  0
#> 131 261_263  0
#> 132 263_265  1
#> 133 265_267  0
#> 134 267_269  1
#> 135 269_271  0
#> 136 271_273  3
#> 137 273_275  0
#> 138 275_277  0
#> 139 277_279  0
#> 140 279_281  1
#> 141 281_283  1
#> 142 283_285  0
#> 143 285_287  1
#> 144 287_289  0
#> 145 289_291  1
#> 146 291_293  1
#> 147 293_295  1
#> 148 295_297  0
#> 149 297_299  0
#> 150 299_301  0
#> 151 301_303  2
#> 152 303_305  1
#> 153 305_307  1
#> 154 307_309  4
#> 155 309_311  1
#> 156 311_313  3
#> 157 313_315  2
#> 158 315_317  3
#> 159 317_319  7
#> 160 319_321  1
#> 161 321_323  9
#> 162 323_325  2
#> 163 325_327  4
#> 164 327_329  2
#> 165 329_331  4
#> 166 331_333  2
#> 167 333_335  1
#> 168 335_337  3
#> 169 337_339  2
#> 170 339_341  3
#> 171 341_343  2
#> 172 343_345  5
#> 173 345_347  3
#> 174 347_349  1
#> 175 349_351  7
#> 176 351_353  4
#> 177 353_355  1
#> 178 355_357  1
#> 179 357_359  1
#> 180 359_361  2
#> 181 361_363  4
#> 182 363_365  7
#> 183 365_367  3
#> 184 367_369  0
#> 185 369_371  3
#> 186 371_373  2
#> 187 373_375  2
#> 188 375_377  4
#> 189 377_379  1
#> 190 379_381  3
#> 191 381_383  2
#> 192 383_385  0
#> 193 385_387  0
#> 194 387_389  0
#> 195 389_391  6
#> 196 391_393  1
#> 197 393_395  0
#> 198 395_397  0
#> 199 397_399  2
#> 200 399_400  0

## with normalization
bin_fragment_size(bam = bamT1, normalized = TRUE)

#>      Breaks           T1
#> 1       1_3 0.0000000000
#> 2       3_5 0.0000000000
#> 3       5_7 0.0000000000
#> 4       7_9 0.0000000000
#> 5      9_11 0.0000000000
#> 6     11_13 0.0000000000
#> 7     13_15 0.0000000000
#> 8     15_17 0.0000000000
#> 9     17_19 0.0000000000
#> 10    19_21 0.0000000000
#> 11    21_23 0.0000000000
#> 12    23_25 0.0000000000
#> 13    25_27 0.0000000000
#> 14    27_29 0.0000000000
#> 15    29_31 0.0000000000
#> 16    31_33 0.0000000000
#> 17    33_35 0.0000000000
#> 18    35_37 0.0000000000
#> 19    37_39 0.0000000000
#> 20    39_41 0.0000000000
#> 21    41_43 0.0000000000
#> 22    43_45 0.0000000000
#> 23    45_47 0.0000000000
#> 24    47_49 0.0000000000
#> 25    49_51 0.0000000000
#> 26    51_53 0.0000000000
#> 27    53_55 0.0000000000
#> 28    55_57 0.0000000000
#> 29    57_59 0.0000000000
#> 30    59_61 0.0000000000
#> 31    61_63 0.0000000000
#> 32    63_65 0.0000000000
#> 33    65_67 0.0000000000
#> 34    67_69 0.0000000000
#> 35    69_71 0.0000000000
#> 36    71_73 0.0000000000
#> 37    73_75 0.0000000000
#> 38    75_77 0.0000000000
#> 39    77_79 0.0000000000
#> 40    79_81 0.0000000000
#> 41    81_83 0.0000000000
#> 42    83_85 0.0000000000
#> 43    85_87 0.0000000000
#> 44    87_89 0.0000000000
#> 45    89_91 0.0000000000
#> 46    91_93 0.0000000000
#> 47    93_95 0.0000000000
#> 48    95_97 0.0000000000
#> 49    97_99 0.0000000000
#> 50   99_101 0.0000000000
#> 51  101_103 0.0000000000
#> 52  103_105 0.0008156607
#> 53  105_107 0.0000000000
#> 54  107_109 0.0000000000
#> 55  109_111 0.0000000000
#> 56  111_113 0.0024469821
#> 57  113_115 0.0000000000
#> 58  115_117 0.0040783034
#> 59  117_119 0.0008156607
#> 60  119_121 0.0024469821
#> 61  121_123 0.0073409462
#> 62  123_125 0.0008156607
#> 63  125_127 0.0008156607
#> 64  127_129 0.0024469821
#> 65  129_131 0.0040783034
#> 66  131_133 0.0073409462
#> 67  133_135 0.0032626427
#> 68  135_137 0.0048939641
#> 69  137_139 0.0057096248
#> 70  139_141 0.0073409462
#> 71  141_143 0.0106035889
#> 72  143_145 0.0122349103
#> 73  145_147 0.0106035889
#> 74  147_149 0.0138662316
#> 75  149_151 0.0228384992
#> 76  151_153 0.0154975530
#> 77  153_155 0.0171288744
#> 78  155_157 0.0195758564
#> 79  157_159 0.0277324633
#> 80  159_161 0.0277324633
#> 81  161_163 0.0244698206
#> 82  163_165 0.0497553018
#> 83  165_167 0.0424143556
#> 84  167_169 0.0522022838
#> 85  169_171 0.0358890701
#> 86  171_173 0.0448613377
#> 87  173_175 0.0318107667
#> 88  175_177 0.0309951060
#> 89  177_179 0.0367047308
#> 90  179_181 0.0350734095
#> 91  181_183 0.0277324633
#> 92  183_185 0.0179445351
#> 93  185_187 0.0220228385
#> 94  187_189 0.0285481240
#> 95  189_191 0.0171288744
#> 96  191_193 0.0114192496
#> 97  193_195 0.0171288744
#> 98  195_197 0.0122349103
#> 99  197_199 0.0130505710
#> 100 199_201 0.0065252855
#> 101 201_203 0.0089722675
#> 102 203_205 0.0065252855
#> 103 205_207 0.0032626427
#> 104 207_209 0.0073409462
#> 105 209_211 0.0032626427
#> 106 211_213 0.0106035889
#> 107 213_215 0.0057096248
#> 108 215_217 0.0040783034
#> 109 217_219 0.0057096248
#> 110 219_221 0.0040783034
#> 111 221_223 0.0032626427
#> 112 223_225 0.0032626427
#> 113 225_227 0.0040783034
#> 114 227_229 0.0024469821
#> 115 229_231 0.0008156607
#> 116 231_233 0.0024469821
#> 117 233_235 0.0032626427
#> 118 235_237 0.0040783034
#> 119 237_239 0.0008156607
#> 120 239_241 0.0040783034
#> 121 241_243 0.0008156607
#> 122 243_245 0.0016313214
#> 123 245_247 0.0024469821
#> 124 247_249 0.0016313214
#> 125 249_251 0.0008156607
#> 126 251_253 0.0016313214
#> 127 253_255 0.0000000000
#> 128 255_257 0.0008156607
#> 129 257_259 0.0016313214
#> 130 259_261 0.0000000000
#> 131 261_263 0.0000000000
#> 132 263_265 0.0008156607
#> 133 265_267 0.0000000000
#> 134 267_269 0.0008156607
#> 135 269_271 0.0000000000
#> 136 271_273 0.0024469821
#> 137 273_275 0.0000000000
#> 138 275_277 0.0000000000
#> 139 277_279 0.0000000000
#> 140 279_281 0.0008156607
#> 141 281_283 0.0008156607
#> 142 283_285 0.0000000000
#> 143 285_287 0.0008156607
#> 144 287_289 0.0000000000
#> 145 289_291 0.0008156607
#> 146 291_293 0.0008156607
#> 147 293_295 0.0008156607
#> 148 295_297 0.0000000000
#> 149 297_299 0.0000000000
#> 150 299_301 0.0000000000
#> 151 301_303 0.0016313214
#> 152 303_305 0.0008156607
#> 153 305_307 0.0008156607
#> 154 307_309 0.0032626427
#> 155 309_311 0.0008156607
#> 156 311_313 0.0024469821
#> 157 313_315 0.0016313214
#> 158 315_317 0.0024469821
#> 159 317_319 0.0057096248
#> 160 319_321 0.0008156607
#> 161 321_323 0.0073409462
#> 162 323_325 0.0016313214
#> 163 325_327 0.0032626427
#> 164 327_329 0.0016313214
#> 165 329_331 0.0032626427
#> 166 331_333 0.0016313214
#> 167 333_335 0.0008156607
#> 168 335_337 0.0024469821
#> 169 337_339 0.0016313214
#> 170 339_341 0.0024469821
#> 171 341_343 0.0016313214
#> 172 343_345 0.0040783034
#> 173 345_347 0.0024469821
#> 174 347_349 0.0008156607
#> 175 349_351 0.0057096248
#> 176 351_353 0.0032626427
#> 177 353_355 0.0008156607
#> 178 355_357 0.0008156607
#> 179 357_359 0.0008156607
#> 180 359_361 0.0016313214
#> 181 361_363 0.0032626427
#> 182 363_365 0.0057096248
#> 183 365_367 0.0024469821
#> 184 367_369 0.0000000000
#> 185 369_371 0.0024469821
#> 186 371_373 0.0016313214
#> 187 373_375 0.0016313214
#> 188 375_377 0.0032626427
#> 189 377_379 0.0008156607
#> 190 379_381 0.0024469821
#> 191 381_383 0.0016313214
#> 192 383_385 0.0000000000
#> 193 385_387 0.0000000000
#> 194 387_389 0.0000000000
#> 195 389_391 0.0048939641
#> 196 391_393 0.0008156607
#> 197 393_395 0.0000000000
#> 198 395_397 0.0000000000
#> 199 397_399 0.0016313214
#> 200 399_400 0.0000000000

## binning reads categorized based on mutations ref and alt
bin_fragment_size(bam = bamT1, mutations = mutations)

#>      Breaks T1_alt T1_ref T1_other
#> 1       1_3      0      0        0
#> 2       3_5      0      0        0
#> 3       5_7      0      0        0
#> 4       7_9      0      0        0
#> 5      9_11      0      0        0
#> 6     11_13      0      0        0
#> 7     13_15      0      0        0
#> 8     15_17      0      0        0
#> 9     17_19      0      0        0
#> 10    19_21      0      0        0
#> 11    21_23      0      0        0
#> 12    23_25      0      0        0
#> 13    25_27      0      0        0
#> 14    27_29      0      0        0
#> 15    29_31      0      0        0
#> 16    31_33      0      0        0
#> 17    33_35      0      0        0
#> 18    35_37      0      0        0
#> 19    37_39      0      0        0
#> 20    39_41      0      0        0
#> 21    41_43      0      0        0
#> 22    43_45      0      0        0
#> 23    45_47      0      0        0
#> 24    47_49      0      0        0
#> 25    49_51      0      0        0
#> 26    51_53      0      0        0
#> 27    53_55      0      0        0
#> 28    55_57      0      0        0
#> 29    57_59      0      0        0
#> 30    59_61      0      0        0
#> 31    61_63      0      0        0
#> 32    63_65      0      0        0
#> 33    65_67      0      0        0
#> 34    67_69      0      0        0
#> 35    69_71      0      0        0
#> 36    71_73      0      0        0
#> 37    73_75      0      0        0
#> 38    75_77      0      0        0
#> 39    77_79      0      0        0
#> 40    79_81      0      0        0
#> 41    81_83      0      0        0
#> 42    83_85      0      0        0
#> 43    85_87      0      0        0
#> 44    87_89      0      0        0
#> 45    89_91      0      0        0
#> 46    91_93      0      0        0
#> 47    93_95      0      0        0
#> 48    95_97      0      0        0
#> 49    97_99      0      0        0
#> 50   99_101      0      0        0
#> 51  101_103      0      0        0
#> 52  103_105      0      1        0
#> 53  105_107      0      0        0
#> 54  107_109      0      0        0
#> 55  109_111      0      0        0
#> 56  111_113      0      3        0
#> 57  113_115      0      0        0
#> 58  115_117      0      3        2
#> 59  117_119      0      1        0
#> 60  119_121      0      2        1
#> 61  121_123      0      8        1
#> 62  123_125      0      1        0
#> 63  125_127      0      1        0
#> 64  127_129      0      2        1
#> 65  129_131      0      5        0
#> 66  131_133      0      9        0
#> 67  133_135      0      3        1
#> 68  135_137      0      6        0
#> 69  137_139      0      5        2
#> 70  139_141      0      8        1
#> 71  141_143      0     13        0
#> 72  143_145      0     15        0
#> 73  145_147      0     12        1
#> 74  147_149      0     17        0
#> 75  149_151      0     26        2
#> 76  151_153      0     18        1
#> 77  153_155      0     20        1
#> 78  155_157      0     22        2
#> 79  157_159      0     33        1
#> 80  159_161      0     33        1
#> 81  161_163      0     29        1
#> 82  163_165      0     58        3
#> 83  165_167      0     51        1
#> 84  167_169      0     62        2
#> 85  169_171      0     43        1
#> 86  171_173      1     51        3
#> 87  173_175      0     39        0
#> 88  175_177      0     38        0
#> 89  177_179      0     44        1
#> 90  179_181      0     43        0
#> 91  181_183      0     33        1
#> 92  183_185      0     21        1
#> 93  185_187      0     27        0
#> 94  187_189      0     35        0
#> 95  189_191      0     21        0
#> 96  191_193      0     14        0
#> 97  193_195      0     21        0
#> 98  195_197      0     15        0
#> 99  197_199      0     16        0
#> 100 199_201      0      8        0
#> 101 201_203      0     11        0
#> 102 203_205      0      8        0
#> 103 205_207      0      4        0
#> 104 207_209      0      8        1
#> 105 209_211      0      4        0
#> 106 211_213      0     12        1
#> 107 213_215      0      7        0
#> 108 215_217      0      5        0
#> 109 217_219      1      6        0
#> 110 219_221      0      4        1
#> 111 221_223      0      4        0
#> 112 223_225      0      3        1
#> 113 225_227      1      4        0
#> 114 227_229      1      2        0
#> 115 229_231      0      0        1
#> 116 231_233      0      3        0
#> 117 233_235      0      4        0
#> 118 235_237      0      5        0
#> 119 237_239      0      1        0
#> 120 239_241      0      5        0
#> 121 241_243      0      1        0
#> 122 243_245      0      2        0
#> 123 245_247      0      3        0
#> 124 247_249      1      0        1
#> 125 249_251      0      1        0
#> 126 251_253      0      2        0
#> 127 253_255      0      0        0
#> 128 255_257      0      1        0
#> 129 257_259      0      2        0
#> 130 259_261      0      0        0
#> 131 261_263      0      0        0
#> 132 263_265      0      0        1
#> 133 265_267      0      0        0
#> 134 267_269      0      1        0
#> 135 269_271      0      0        0
#> 136 271_273      0      3        0
#> 137 273_275      0      0        0
#> 138 275_277      0      0        0
#> 139 277_279      0      0        0
#> 140 279_281      0      1        0
#> 141 281_283      0      1        0
#> 142 283_285      0      0        0
#> 143 285_287      0      0        1
#> 144 287_289      0      0        0
#> 145 289_291      0      1        0
#> 146 291_293      0      1        0
#> 147 293_295      0      1        0
#> 148 295_297      0      0        0
#> 149 297_299      0      0        0
#> 150 299_301      0      0        0
#> 151 301_303      0      2        0
#> 152 303_305      0      1        0
#> 153 305_307      0      1        0
#> 154 307_309      0      4        0
#> 155 309_311      0      1        0
#> 156 311_313      0      3        0
#> 157 313_315      0      2        0
#> 158 315_317      0      3        0
#> 159 317_319      0      7        0
#> 160 319_321      0      1        0
#> 161 321_323      0      9        0
#> 162 323_325      0      2        0
#> 163 325_327      0      4        0
#> 164 327_329      0      2        0
#> 165 329_331      0      4        0
#> 166 331_333      0      2        0
#> 167 333_335      0      1        0
#> 168 335_337      0      3        0
#> 169 337_339      0      2        0
#> 170 339_341      0      3        0
#> 171 341_343      0      2        0
#> 172 343_345      0      5        0
#> 173 345_347      0      3        0
#> 174 347_349      0      1        0
#> 175 349_351      0      7        0
#> 176 351_353      0      4        0
#> 177 353_355      0      1        0
#> 178 355_357      0      1        0
#> 179 357_359      0      1        0
#> 180 359_361      0      2        0
#> 181 361_363      1      3        0
#> 182 363_365      0      7        0
#> 183 365_367      0      3        0
#> 184 367_369      0      0        0
#> 185 369_371      0      3        0
#> 186 371_373      0      2        0
#> 187 373_375      0      2        0
#> 188 375_377      0      4        0
#> 189 377_379      0      1        0
#> 190 379_381      0      3        0
#> 191 381_383      0      2        0
#> 192 383_385      0      0        0
#> 193 385_387      0      0        0
#> 194 387_389      0      0        0
#> 195 389_391      0      6        0
#> 196 391_393      0      1        0
#> 197 393_395      0      0        0
#> 198 395_397      0      0        0
#> 199 397_399      0      2        0
#> 200 399_400      0      0        0

## Restrict to reads into targets
bin_fragment_size(bam = bamT1, targets = targets)

#>      Breaks T1
#> 1       1_3  0
#> 2       3_5  0
#> 3       5_7  0
#> 4       7_9  0
#> 5      9_11  0
#> 6     11_13  0
#> 7     13_15  0
#> 8     15_17  0
#> 9     17_19  0
#> 10    19_21  0
#> 11    21_23  0
#> 12    23_25  0
#> 13    25_27  0
#> 14    27_29  0
#> 15    29_31  0
#> 16    31_33  0
#> 17    33_35  0
#> 18    35_37  0
#> 19    37_39  0
#> 20    39_41  0
#> 21    41_43  0
#> 22    43_45  0
#> 23    45_47  0
#> 24    47_49  0
#> 25    49_51  0
#> 26    51_53  0
#> 27    53_55  0
#> 28    55_57  0
#> 29    57_59  0
#> 30    59_61  0
#> 31    61_63  0
#> 32    63_65  0
#> 33    65_67  0
#> 34    67_69  0
#> 35    69_71  0
#> 36    71_73  0
#> 37    73_75  0
#> 38    75_77  0
#> 39    77_79  0
#> 40    79_81  0
#> 41    81_83  0
#> 42    83_85  0
#> 43    85_87  0
#> 44    87_89  0
#> 45    89_91  0
#> 46    91_93  0
#> 47    93_95  0
#> 48    95_97  0
#> 49    97_99  0
#> 50   99_101  0
#> 51  101_103  0
#> 52  103_105  1
#> 53  105_107  0
#> 54  107_109  0
#> 55  109_111  0
#> 56  111_113  3
#> 57  113_115  0
#> 58  115_117  5
#> 59  117_119  1
#> 60  119_121  3
#> 61  121_123  9
#> 62  123_125  1
#> 63  125_127  1
#> 64  127_129  3
#> 65  129_131  5
#> 66  131_133  9
#> 67  133_135  4
#> 68  135_137  6
#> 69  137_139  7
#> 70  139_141  9
#> 71  141_143 13
#> 72  143_145 15
#> 73  145_147 13
#> 74  147_149 17
#> 75  149_151 28
#> 76  151_153 19
#> 77  153_155 21
#> 78  155_157 24
#> 79  157_159 34
#> 80  159_161 34
#> 81  161_163 30
#> 82  163_165 61
#> 83  165_167 52
#> 84  167_169 64
#> 85  169_171 44
#> 86  171_173 55
#> 87  173_175 39
#> 88  175_177 38
#> 89  177_179 45
#> 90  179_181 43
#> 91  181_183 34
#> 92  183_185 22
#> 93  185_187 27
#> 94  187_189 35
#> 95  189_191 21
#> 96  191_193 14
#> 97  193_195 21
#> 98  195_197 15
#> 99  197_199 16
#> 100 199_201  8
#> 101 201_203 11
#> 102 203_205  8
#> 103 205_207  4
#> 104 207_209  9
#> 105 209_211  4
#> 106 211_213 13
#> 107 213_215  7
#> 108 215_217  5
#> 109 217_219  7
#> 110 219_221  5
#> 111 221_223  4
#> 112 223_225  4
#> 113 225_227  5
#> 114 227_229  3
#> 115 229_231  1
#> 116 231_233  3
#> 117 233_235  4
#> 118 235_237  5
#> 119 237_239  1
#> 120 239_241  5
#> 121 241_243  1
#> 122 243_245  2
#> 123 245_247  3
#> 124 247_249  2
#> 125 249_251  1
#> 126 251_253  2
#> 127 253_255  0
#> 128 255_257  1
#> 129 257_259  2
#> 130 259_261  0
#> 131 261_263  0
#> 132 263_265  1
#> 133 265_267  0
#> 134 267_269  1
#> 135 269_271  0
#> 136 271_273  3
#> 137 273_275  0
#> 138 275_277  0
#> 139 277_279  0
#> 140 279_281  1
#> 141 281_283  1
#> 142 283_285  0
#> 143 285_287  1
#> 144 287_289  0
#> 145 289_291  1
#> 146 291_293  1
#> 147 293_295  1
#> 148 295_297  0
#> 149 297_299  0
#> 150 299_301  0
#> 151 301_303  2
#> 152 303_305  1
#> 153 305_307  1
#> 154 307_309  4
#> 155 309_311  1
#> 156 311_313  3
#> 157 313_315  2
#> 158 315_317  3
#> 159 317_319  7
#> 160 319_321  1
#> 161 321_323  9
#> 162 323_325  2
#> 163 325_327  4
#> 164 327_329  2
#> 165 329_331  4
#> 166 331_333  2
#> 167 333_335  1
#> 168 335_337  3
#> 169 337_339  2
#> 170 339_341  3
#> 171 341_343  2
#> 172 343_345  5
#> 173 345_347  3
#> 174 347_349  1
#> 175 349_351  7
#> 176 351_353  4
#> 177 353_355  1
#> 178 355_357  1
#> 179 357_359  1
#> 180 359_361  2
#> 181 361_363  4
#> 182 363_365  7
#> 183 365_367  3
#> 184 367_369  0
#> 185 369_371  3
#> 186 371_373  2
#> 187 373_375  2
#> 188 375_377  4
#> 189 377_379  1
#> 190 379_381  3
#> 191 381_383  2
#> 192 383_385  0
#> 193 385_387  0
#> 194 387_389  0
#> 195 389_391  6
#> 196 391_393  1
#> 197 393_395  0
#> 198 395_397  0
#> 199 397_399  2
#> 200 399_400  0

# }