Create a background panel instance from a bam file (e.g. healthy samples)

This function scans the targets regions in one bam file, and reports the number of reference, non-reference reads for each loci in addition to the non-reference (VAF) allele frequency. Loci with VAF higher than vaf_threshold are masked with NA. This function is used internally by create_background_panel

create_background_panel_instance(bam, targets, reference,
  vaf_threshold = 0.05, tag = "", min_base_quality = 20,
  max_depth = 1e+05, min_mapq = 30, substitution_specific = TRUE)

Arguments

bam
A character specifying the path to bam file.

targets
The targets data frame must have the columns chr, start and end.

reference
The reference genome as BSgenome object.

vaf_threshold
Loci with the fraction of non-reference reads above this value are masked with NA.

tag
The RG tag in the bam file. Empty string defaults to using the whole bam.

min_base_quality
The minimum base quality to count a read for a loci.

max_depth
Maximum depth for the pileup

min_mapq
The minimum mapping quality to count a read for a loci

substitution_specific
logical, whether to have the loci by substitutions.

Value

A named list having depth, alt and vaf data frames. Each has the same order of loci in rows and the input sample in columns.