Get names of the reads showing alternative allele of a mutation

Extract the names of the reads in a bam file that support the variant and reference alleles of the input mutations

get_mutations_read_names(bam, mutations, min_base_quality = 20,
  tag = "", min_mapq = 30)

Arguments

bam
path to bam file

mutations
A data frame containing the mutations. Must have the columns CHROM, POS, REF, ALT.

min_base_quality
integer specifying the minimum base quality for reads to be included.

tag
the RG tag if the bam has more than one samplee

min_mapq
integer specifying the minimum mapping quality for reads to be included

Value

A list with length equal to the number of mutations. Each element is a character vector with the read names.

Details

Returns the IDs of the read that cover the input mutations (ref and alt alleles).

See also

Examples

data('mutations',package = 'ctDNAtools')
bamT1 <- system.file('extdata', 'T1.bam', package = 'ctDNAtools')
get_mutations_read_names(bam = bamT1, mutations = mutations[1:3,])

#> $`chr14:106327474_C_G`
#> $`chr14:106327474_C_G`$ref
#>   [1] "T1_Library1:1954851" "T1_Library1:1955625" "T1_Library1:1956206"
#>   [4] "T1_Library1:1956604" "T1_Library1:1956608" "T1_Library1:1956615"
#>   [7] "T1_Library1:1956618" "T1_Library1:1956313" "T1_Library1:1956619"
#>  [10] "T1_Library1:1954604" "T1_Library1:1956103" "T1_Library1:1956157"
#>  [13] "T1_Library1:1956626" "T1_Library1:1956167" "T1_Library1:1956628"
#>  [16] "T1_Library1:1956631" "T1_Library1:1956081" "T1_Library1:1956186"
#>  [19] "T1_Library1:1956634" "T1_Library1:1956138" "T1_Library1:1956170"
#>  [22] "T1_Library1:1956368" "T1_Library1:1956637" "T1_Library1:1956639"
#>  [25] "T1_Library1:1956640" "T1_Library1:1956641" "T1_Library1:1956643"
#>  [28] "T1_Library1:1956644" "T1_Library1:1956646" "T1_Library1:1956649"
#>  [31] "T1_Library1:1956650" "T1_Library1:1956651" "T1_Library1:1956655"
#>  [34] "T1_Library1:1956662" "T1_Library1:1956664" "T1_Library1:1956666"
#>  [37] "T1_Library1:1954235" "T1_Library1:1954993" "T1_Library1:1955012"
#>  [40] "T1_Library1:1955454" "T1_Library1:1955614" "T1_Library1:1956132"
#>  [43] "T1_Library1:1956141" "T1_Library1:1956142" "T1_Library1:1956158"
#>  [46] "T1_Library1:1956244" "T1_Library1:1956246" "T1_Library1:1956667"
#>  [49] "T1_Library1:1956668" "T1_Library1:1956669" "T1_Library1:1956670"
#>  [52] "T1_Library1:1956672" "T1_Library1:1956677" "T1_Library1:1956679"
#>  [55] "T1_Library1:1956680" "T1_Library1:1956686" "T1_Library1:1956687"
#>  [58] "T1_Library1:1956688" "T1_Library1:1956694" "T1_Library1:1956695"
#>  [61] "T1_Library1:1956696" "T1_Library1:1956698" "T1_Library1:1956701"
#>  [64] "T1_Library1:1956703" "T1_Library1:1956715" "T1_Library1:1956716"
#>  [67] "T1_Library1:1954887" "T1_Library1:1956150" "T1_Library1:1956718"
#>  [70] "T1_Library1:1956719" "T1_Library1:1956155" "T1_Library1:1956720"
#>  [73] "T1_Library1:1954822" "T1_Library1:1956727" "T1_Library1:1956731"
#>  [76] "T1_Library1:1955221" "T1_Library1:1956109" "T1_Library1:1956110"
#>  [79] "T1_Library1:1956121" "T1_Library1:1956127" "T1_Library1:1956218"
#>  [82] "T1_Library1:1956743" "T1_Library1:1955222" "T1_Library1:1956152"
#>  [85] "T1_Library1:1956153" "T1_Library1:1956159" "T1_Library1:1956211"
#>  [88] "T1_Library1:1956372" "T1_Library1:1956373" "T1_Library1:1956746"
#>  [91] "T1_Library1:1956749" "T1_Library1:1956751" "T1_Library1:1956755"
#>  [94] "T1_Library1:1956188" "T1_Library1:1956757" "T1_Library1:1956123"
#>  [97] "T1_Library1:1956760" "T1_Library1:1956762" "T1_Library1:1956766"
#> [100] "T1_Library1:1956768" "T1_Library1:1956773" "T1_Library1:1956212"
#> [103] "T1_Library1:1956249" "T1_Library1:1956250" "T1_Library1:1956780"
#> [106] "T1_Library1:1956784" "T1_Library1:1956786" "T1_Library1:1956788"
#> [109] "T1_Library1:1956796" "T1_Library1:1956798" "T1_Library1:1956804"
#> [112] "T1_Library1:1956805" "T1_Library1:1956806" "T1_Library1:1956807"
#> [115] "T1_Library1:1956810" "T1_Library1:1956814" "T1_Library1:1956819"
#> [118] "T1_Library1:1956822" "T1_Library1:1956823" "T1_Library1:1956825"
#> [121] "T1_Library1:1956829" "T1_Library1:1956831" "T1_Library1:1956836"
#> [124] "T1_Library1:1956307" "T1_Library1:1956846" "T1_Library1:1956848"
#> [127] "T1_Library1:1956028" "T1_Library1:1956129" "T1_Library1:1956191"
#> [130] "T1_Library1:1956194" "T1_Library1:1956857" "T1_Library1:1955876"
#> [133] "T1_Library1:1956862" "T1_Library1:1956866" "T1_Library1:1956868"
#> [136] "T1_Library1:1956871" "T1_Library1:1956875" "T1_Library1:1956878"
#> [139] "T1_Library1:1956879" "T1_Library1:1956880" "T1_Library1:1956881"
#> [142] "T1_Library1:1954624" "T1_Library1:1956229" "T1_Library1:1956336"
#> [145] "T1_Library1:1955509" "T1_Library1:1956292" "T1_Library1:1956252"
#> [148] "T1_Library1:1956311" "T1_Library1:1956891" "T1_Library1:1956894"
#> [151] "T1_Library1:1956255" "T1_Library1:1956896" "T1_Library1:1956897"
#> [154] "T1_Library1:1956900" "T1_Library1:1956906" "T1_Library1:1956907"
#> [157] "T1_Library1:1955748" "T1_Library1:1956910" "T1_Library1:1956911"
#> [160] "T1_Library1:1956916" "T1_Library1:1956917" "T1_Library1:1956918"
#> [163] "T1_Library1:1956922" "T1_Library1:1956924" "T1_Library1:1956926"
#> [166] "T1_Library1:1956927" "T1_Library1:1956930" "T1_Library1:1956931"
#> [169] "T1_Library1:1956937" "T1_Library1:1956940" "T1_Library1:1956941"
#> [172] "T1_Library1:1956945" "T1_Library1:1952419" "T1_Library1:1956331"
#> [175] "T1_Library1:1956400" "T1_Library1:1956448" "T1_Library1:1956469"
#> [178] "T1_Library1:1956958" "T1_Library1:1955049" "T1_Library1:1956375"
#> [181] "T1_Library1:1956483" "T1_Library1:1956144" "T1_Library1:1956178"
#> [184] "T1_Library1:1956356" "T1_Library1:1956485" "T1_Library1:1956965"
#> [187] "T1_Library1:1956261" "T1_Library1:1956435" "T1_Library1:1956966"
#> [190] "T1_Library1:1956967" "T1_Library1:1955366" "T1_Library1:1956317"
#> [193] "T1_Library1:1956977" "T1_Library1:1956978" "T1_Library1:1956263"
#> [196] "T1_Library1:1956538" "T1_Library1:1956982" "T1_Library1:1956989"
#> [199] "T1_Library1:1956990" "T1_Library1:1956992" "T1_Library1:1955056"
#> [202] "T1_Library1:1956226" "T1_Library1:1956402" "T1_Library1:1956057"
#> [205] "T1_Library1:1953783" "T1_Library1:1955615" "T1_Library1:1956267"
#> [208] "T1_Library1:1956269" "T1_Library1:1956299" "T1_Library1:1956344"
#> [211] "T1_Library1:1956363" "T1_Library1:1956394" "T1_Library1:1956432"
#> [214] "T1_Library1:1956471" "T1_Library1:1956519" "T1_Library1:1956520"
#> [217] "T1_Library1:1956539" "T1_Library1:1956558" "T1_Library1:1957011"
#> [220] "T1_Library1:1957013" "T1_Library1:1957014" "T1_Library1:1957016"
#> [223] "T1_Library1:1957022" "T1_Library1:1957024" "T1_Library1:1957030"
#> [226] "T1_Library1:1957031" "T1_Library1:1957035" "T1_Library1:1957039"
#> [229] "T1_Library1:1957040" "T1_Library1:1957046" "T1_Library1:1957047"
#> [232] "T1_Library1:1957048" "T1_Library1:1956233" "T1_Library1:1956339"
#> [235] "T1_Library1:1957050" "T1_Library1:1957052" "T1_Library1:1956380"
#> [238] "T1_Library1:1957054" "T1_Library1:1957055" "T1_Library1:1956403"
#> [241] "T1_Library1:1957060" "T1_Library1:1957061" "T1_Library1:1955229"
#> [244] "T1_Library1:1955728" "T1_Library1:1955843" "T1_Library1:1956272"
#> [247] "T1_Library1:1956293" "T1_Library1:1956321" "T1_Library1:1956437"
#> [250] "T1_Library1:1956445" "T1_Library1:1956456" "T1_Library1:1956541"
#> [253] "T1_Library1:1956559" "T1_Library1:1957065" "T1_Library1:1955480"
#> [256] "T1_Library1:1956581" "T1_Library1:1956582" "T1_Library1:1957067"
#> [259] "T1_Library1:1957076" "T1_Library1:1957077" "T1_Library1:1957079"
#> [262] "T1_Library1:1957080" "T1_Library1:1957084" "T1_Library1:1957087"
#> [265] "T1_Library1:1957092" "T1_Library1:1955026" "T1_Library1:1956439"
#> [268] "T1_Library1:1956525" "T1_Library1:1956562" "T1_Library1:1957093"
#> [271] "T1_Library1:1957094" "T1_Library1:1957097" "T1_Library1:1957099"
#> [274] "T1_Library1:1956413" "T1_Library1:1956502" "T1_Library1:1957102"
#> [277] "T1_Library1:1957104" "T1_Library1:1957105" "T1_Library1:1957106"
#> [280] "T1_Library1:1957107" "T1_Library1:1955230" "T1_Library1:1956405"
#> [283] "T1_Library1:1956415" "T1_Library1:1956477" "T1_Library1:1954863"
#> [286] "T1_Library1:1956199" "T1_Library1:1957117" "T1_Library1:1957124"
#> [289] "T1_Library1:1957125" "T1_Library1:1955729" "T1_Library1:1955645"
#> [292] "T1_Library1:1957130" "T1_Library1:1957132" "T1_Library1:1957136"
#> [295] "T1_Library1:1957137" "T1_Library1:1957140" "T1_Library1:1957141"
#> [298] "T1_Library1:1957145" "T1_Library1:1957147" "T1_Library1:1957148"
#> [301] "T1_Library1:1955590" "T1_Library1:1957152" "T1_Library1:1957153"
#> [304] "T1_Library1:1955923" "T1_Library1:1956029" "T1_Library1:1956030"
#> [307] "T1_Library1:1956113" "T1_Library1:1956365" "T1_Library1:1957157"
#> [310] "T1_Library1:1957159" "T1_Library1:1957160" "T1_Library1:1957161"
#> [313] "T1_Library1:1957165" "T1_Library1:1957166" "T1_Library1:1957167"
#> [316] "T1_Library1:1957168" "T1_Library1:1956564" "T1_Library1:1957172"
#> [319] "T1_Library1:1957174" "T1_Library1:1955485" "T1_Library1:1955845"
#> [322] "T1_Library1:1956440" "T1_Library1:1956452" "T1_Library1:1956458"
#> [325] "T1_Library1:1956276" "T1_Library1:1956490" "T1_Library1:1957177"
#> [328] "T1_Library1:1957179" "T1_Library1:1957182" "T1_Library1:1955791"
#> [331] "T1_Library1:1957185" "T1_Library1:1956527" "T1_Library1:1957190"
#> [334] "T1_Library1:1957198" "T1_Library1:1957199" "T1_Library1:1957200"
#> [337] "T1_Library1:1956419" "T1_Library1:1957201" "T1_Library1:1955413"
#> [340] "T1_Library1:1956360" "T1_Library1:1953517" "T1_Library1:1955913"
#> [343] "T1_Library1:1956517" "T1_Library1:1955977" "T1_Library1:1956077"
#> [346] "T1_Library1:1956522" "T1_Library1:1956200" "T1_Library1:1956585"
#> [349] "T1_Library1:1956528" "T1_Library1:1955287" "T1_Library1:1957225"
#> [352] "T1_Library1:1957226" "T1_Library1:1957227" "T1_Library1:1956586"
#> [355] "T1_Library1:1956407" "T1_Library1:1957235" "T1_Library1:1957239"
#> [358] "T1_Library1:1957240" "T1_Library1:1956351" "T1_Library1:1956479"
#> [361] "T1_Library1:1957244" "T1_Library1:1955733" "T1_Library1:1957248"
#> [364] "T1_Library1:1955620" "T1_Library1:1956215" "T1_Library1:1956425"
#> [367] "T1_Library1:1955527" "T1_Library1:1956352" "T1_Library1:1957255"
#> [370] "T1_Library1:1955528" "T1_Library1:1956422" "T1_Library1:1957256"
#> [373] "T1_Library1:1957257" "T1_Library1:1957258" "T1_Library1:1957262"
#> 
#> $`chr14:106327474_C_G`$alt
#> [1] "T1_Library1:1956883" "T1_Library1:1957254"
#> 
#> 
#> $`chr14:106327649_G_T`
#> $`chr14:106327649_G_T`$ref
#>   [1] "T1_Library1:1957616" "T1_Library1:1957623" "T1_Library1:1957624"
#>   [4] "T1_Library1:1957636" "T1_Library1:1957638" "T1_Library1:1957640"
#>   [7] "T1_Library1:1957642" "T1_Library1:1957644" "T1_Library1:1957646"
#>  [10] "T1_Library1:1957651" "T1_Library1:1957653" "T1_Library1:1957657"
#>  [13] "T1_Library1:1957658" "T1_Library1:1957660" "T1_Library1:1957661"
#>  [16] "T1_Library1:1957662" "T1_Library1:1957663" "T1_Library1:1957666"
#>  [19] "T1_Library1:1957667" "T1_Library1:1957673" "T1_Library1:1955987"
#>  [22] "T1_Library1:1957675" "T1_Library1:1957677" "T1_Library1:1957679"
#>  [25] "T1_Library1:1957685" "T1_Library1:1957687" "T1_Library1:1957694"
#>  [28] "T1_Library1:1957699" "T1_Library1:1957700" "T1_Library1:1957701"
#>  [31] "T1_Library1:1957703" "T1_Library1:1957713" "T1_Library1:1957715"
#>  [34] "T1_Library1:1957716" "T1_Library1:1957503" "T1_Library1:1957720"
#>  [37] "T1_Library1:1957722" "T1_Library1:1957283" "T1_Library1:1957324"
#>  [40] "T1_Library1:1957731" "T1_Library1:1957738" "T1_Library1:1956281"
#>  [43] "T1_Library1:1956283" "T1_Library1:1956284" "T1_Library1:1956387"
#>  [46] "T1_Library1:1957153" "T1_Library1:1957200" "T1_Library1:1957244"
#>  [49] "T1_Library1:1957276" "T1_Library1:1957284" "T1_Library1:1957286"
#>  [52] "T1_Library1:1957290" "T1_Library1:1957291" "T1_Library1:1957292"
#>  [55] "T1_Library1:1957325" "T1_Library1:1955344" "T1_Library1:1956989"
#>  [58] "T1_Library1:1956990" "T1_Library1:1957288" "T1_Library1:1957336"
#>  [61] "T1_Library1:1957444" "T1_Library1:1957746" "T1_Library1:1957748"
#>  [64] "T1_Library1:1957749" "T1_Library1:1957750" "T1_Library1:1957752"
#>  [67] "T1_Library1:1957754" "T1_Library1:1957762" "T1_Library1:1957766"
#>  [70] "T1_Library1:1957768" "T1_Library1:1957770" "T1_Library1:1957277"
#>  [73] "T1_Library1:1957778" "T1_Library1:1957779" "T1_Library1:1957780"
#>  [76] "T1_Library1:1957781" "T1_Library1:1957788" "T1_Library1:1957789"
#>  [79] "T1_Library1:1957790" "T1_Library1:1957796" "T1_Library1:1957798"
#>  [82] "T1_Library1:1957799" "T1_Library1:1957800" "T1_Library1:1957805"
#>  [85] "T1_Library1:1957807" "T1_Library1:1957815" "T1_Library1:1957823"
#>  [88] "T1_Library1:1957826" "T1_Library1:1957331" "T1_Library1:1957832"
#>  [91] "T1_Library1:1957836" "T1_Library1:1956235" "T1_Library1:1956341"
#>  [94] "T1_Library1:1956463" "T1_Library1:1956977" "T1_Library1:1956978"
#>  [97] "T1_Library1:1957255" "T1_Library1:1957258" "T1_Library1:1957300"
#> [100] "T1_Library1:1956992" "T1_Library1:1957846" "T1_Library1:1957848"
#> [103] "T1_Library1:1957851" "T1_Library1:1957858" "T1_Library1:1957867"
#> [106] "T1_Library1:1957868" "T1_Library1:1957870" "T1_Library1:1957871"
#> [109] "T1_Library1:1957872" "T1_Library1:1957873" "T1_Library1:1954685"
#> [112] "T1_Library1:1957343" "T1_Library1:1957363" "T1_Library1:1957881"
#> [115] "T1_Library1:1957883" "T1_Library1:1957884" "T1_Library1:1957888"
#> [118] "T1_Library1:1957889" "T1_Library1:1957890" "T1_Library1:1957892"
#> [121] "T1_Library1:1957345" "T1_Library1:1957898" "T1_Library1:1957901"
#> [124] "T1_Library1:1957903" "T1_Library1:1957909" "T1_Library1:1957910"
#> [127] "T1_Library1:1957914" "T1_Library1:1957915" "T1_Library1:1957916"
#> [130] "T1_Library1:1957918" "T1_Library1:1957374" "T1_Library1:1957923"
#> [133] "T1_Library1:1957925" "T1_Library1:1957239" "T1_Library1:1957364"
#> [136] "T1_Library1:1957365" "T1_Library1:1957927" "T1_Library1:1957933"
#> [139] "T1_Library1:1957938" "T1_Library1:1957580" "T1_Library1:1957941"
#> [142] "T1_Library1:1957945" "T1_Library1:1957949" "T1_Library1:1957950"
#> [145] "T1_Library1:1957955" "T1_Library1:1957957" "T1_Library1:1957970"
#> [148] "T1_Library1:1957971" "T1_Library1:1957974" "T1_Library1:1957979"
#> [151] "T1_Library1:1957980" "T1_Library1:1957982" "T1_Library1:1957986"
#> [154] "T1_Library1:1957987" "T1_Library1:1957990" "T1_Library1:1957993"
#> [157] "T1_Library1:1957997" "T1_Library1:1957998" "T1_Library1:1958004"
#> [160] "T1_Library1:1958008" "T1_Library1:1958010" "T1_Library1:1958011"
#> [163] "T1_Library1:1957498" "T1_Library1:1958012" "T1_Library1:1958013"
#> [166] "T1_Library1:1958014" "T1_Library1:1958023" "T1_Library1:1957417"
#> [169] "T1_Library1:1958027" "T1_Library1:1958033" "T1_Library1:1958034"
#> [172] "T1_Library1:1958036" "T1_Library1:1958037" "T1_Library1:1958044"
#> [175] "T1_Library1:1958047" "T1_Library1:1958051" "T1_Library1:1958053"
#> [178] "T1_Library1:1958054" "T1_Library1:1958058" "T1_Library1:1958061"
#> [181] "T1_Library1:1958066" "T1_Library1:1958069" "T1_Library1:1958070"
#> [184] "T1_Library1:1958074" "T1_Library1:1958077" "T1_Library1:1958079"
#> [187] "T1_Library1:1958084" "T1_Library1:1958086" "T1_Library1:1956618"
#> [190] "T1_Library1:1957421" "T1_Library1:1957422" "T1_Library1:1957473"
#> [193] "T1_Library1:1957568" "T1_Library1:1958099" "T1_Library1:1958102"
#> [196] "T1_Library1:1958108" "T1_Library1:1958110" "T1_Library1:1957452"
#> [199] "T1_Library1:1957463" "T1_Library1:1958114" "T1_Library1:1958117"
#> [202] "T1_Library1:1958118" "T1_Library1:1958120" "T1_Library1:1958122"
#> [205] "T1_Library1:1956201" "T1_Library1:1956573" "T1_Library1:1957293"
#> [208] "T1_Library1:1957367" "T1_Library1:1957396" "T1_Library1:1957404"
#> [211] "T1_Library1:1958132" "T1_Library1:1958136" "T1_Library1:1958144"
#> [214] "T1_Library1:1958146" "T1_Library1:1958148" "T1_Library1:1958152"
#> [217] "T1_Library1:1957564" "T1_Library1:1958163" "T1_Library1:1958164"
#> [220] "T1_Library1:1958166" "T1_Library1:1958169" "T1_Library1:1958170"
#> [223] "T1_Library1:1958176" "T1_Library1:1958177" "T1_Library1:1958180"
#> [226] "T1_Library1:1958182" "T1_Library1:1958186" "T1_Library1:1958188"
#> [229] "T1_Library1:1958191" "T1_Library1:1958195" "T1_Library1:1958198"
#> [232] "T1_Library1:1958200" "T1_Library1:1958203" "T1_Library1:1957409"
#> [235] "T1_Library1:1957540" "T1_Library1:1958206" "T1_Library1:1957381"
#> [238] "T1_Library1:1957541" "T1_Library1:1958209" "T1_Library1:1958213"
#> [241] "T1_Library1:1958217" "T1_Library1:1958219" "T1_Library1:1958224"
#> [244] "T1_Library1:1958225" "T1_Library1:1956508" "T1_Library1:1957393"
#> [247] "T1_Library1:1957428" "T1_Library1:1957522" "T1_Library1:1957614"
#> [250] "T1_Library1:1958228" "T1_Library1:1957383" "T1_Library1:1957465"
#> [253] "T1_Library1:1957486" "T1_Library1:1957507" "T1_Library1:1957586"
#> [256] "T1_Library1:1958232" "T1_Library1:1958235" "T1_Library1:1958237"
#> [259] "T1_Library1:1957456" "T1_Library1:1957469" "T1_Library1:1957483"
#> [262] "T1_Library1:1958238" "T1_Library1:1958241" "T1_Library1:1958243"
#> [265] "T1_Library1:1958244" "T1_Library1:1958245" "T1_Library1:1957487"
#> [268] "T1_Library1:1958249" "T1_Library1:1958252" "T1_Library1:1958261"
#> [271] "T1_Library1:1957572" "T1_Library1:1957201" "T1_Library1:1958264"
#> [274] "T1_Library1:1958265" "T1_Library1:1958268" "T1_Library1:1958269"
#> [277] "T1_Library1:1958271" "T1_Library1:1958274" "T1_Library1:1958278"
#> [280] "T1_Library1:1958284" "T1_Library1:1958285" "T1_Library1:1958288"
#> [283] "T1_Library1:1958290" "T1_Library1:1958293" "T1_Library1:1958309"
#> [286] "T1_Library1:1958315" "T1_Library1:1958321" "T1_Library1:1958325"
#> [289] "T1_Library1:1956285" "T1_Library1:1956323" "T1_Library1:1956324"
#> [292] "T1_Library1:1958331" "T1_Library1:1955014" "T1_Library1:1955832"
#> [295] "T1_Library1:1956149" "T1_Library1:1957168" "T1_Library1:1957434"
#> [298] "T1_Library1:1957435" "T1_Library1:1957471" "T1_Library1:1957526"
#> [301] "T1_Library1:1957601" "T1_Library1:1958336" "T1_Library1:1958338"
#> [304] "T1_Library1:1955263" "T1_Library1:1958349" "T1_Library1:1958351"
#> [307] "T1_Library1:1958353" "T1_Library1:1957046" "T1_Library1:1957278"
#> [310] "T1_Library1:1957355" "T1_Library1:1957528" "T1_Library1:1958362"
#> [313] "T1_Library1:1958366" "T1_Library1:1958368" "T1_Library1:1958373"
#> [316] "T1_Library1:1958376" "T1_Library1:1956666" "T1_Library1:1956731"
#> [319] "T1_Library1:1958384" "T1_Library1:1956819" "T1_Library1:1957547"
#> [322] "T1_Library1:1958386" "T1_Library1:1958389" "T1_Library1:1958391"
#> [325] "T1_Library1:1958393" "T1_Library1:1956639" "T1_Library1:1957549"
#> [328] "T1_Library1:1957490" "T1_Library1:1958398" "T1_Library1:1956532"
#> [331] "T1_Library1:1958404" "T1_Library1:1956640" "T1_Library1:1956641"
#> [334] "T1_Library1:1957457" "T1_Library1:1957605" "T1_Library1:1958407"
#> [337] "T1_Library1:1956443" "T1_Library1:1958412" "T1_Library1:1958415"
#> [340] "T1_Library1:1958417" "T1_Library1:1958420" "T1_Library1:1956715"
#> [343] "T1_Library1:1956716" "T1_Library1:1958423" "T1_Library1:1958424"
#> [346] "T1_Library1:1958425" "T1_Library1:1958429" "T1_Library1:1957294"
#> [349] "T1_Library1:1957550" "T1_Library1:1958431" "T1_Library1:1958433"
#> [352] "T1_Library1:1957606" "T1_Library1:1958440" "T1_Library1:1958443"
#> [355] "T1_Library1:1958444" "T1_Library1:1958446" "T1_Library1:1957690"
#> [358] "T1_Library1:1958450" "T1_Library1:1958451" "T1_Library1:1954756"
#> [361] "T1_Library1:1958453" "T1_Library1:1958454" "T1_Library1:1958455"
#> [364] "T1_Library1:1958460" "T1_Library1:1958461" "T1_Library1:1958463"
#> [367] "T1_Library1:1958466" "T1_Library1:1958469" "T1_Library1:1956398"
#> 
#> $`chr14:106327649_G_T`$alt
#> [1] "T1_Library1:1956883" "T1_Library1:1957254" "T1_Library1:1958449"
#> 
#> 
#> $`chr14:106327759_A_T`
#> $`chr14:106327759_A_T`$ref
#>   [1] "T1_Library1:1958509" "T1_Library1:1958510" "T1_Library1:1957766"
#>   [4] "T1_Library1:1958512" "T1_Library1:1958514" "T1_Library1:1958517"
#>   [7] "T1_Library1:1957140" "T1_Library1:1957141" "T1_Library1:1958166"
#>  [10] "T1_Library1:1953870" "T1_Library1:1957858" "T1_Library1:1957910"
#>  [13] "T1_Library1:1958519" "T1_Library1:1958520" "T1_Library1:1958525"
#>  [16] "T1_Library1:1958527" "T1_Library1:1958528" "T1_Library1:1958531"
#>  [19] "T1_Library1:1958534" "T1_Library1:1957492" "T1_Library1:1957738"
#>  [22] "T1_Library1:1957867" "T1_Library1:1958004" "T1_Library1:1958132"
#>  [25] "T1_Library1:1956945" "T1_Library1:1957923" "T1_Library1:1958051"
#>  [28] "T1_Library1:1958114" "T1_Library1:1958191" "T1_Library1:1955551"
#>  [31] "T1_Library1:1956511" "T1_Library1:1956757" "T1_Library1:1956773"
#>  [34] "T1_Library1:1957800" "T1_Library1:1957889" "T1_Library1:1957970"
#>  [37] "T1_Library1:1957971" "T1_Library1:1958053" "T1_Library1:1958054"
#>  [40] "T1_Library1:1958077" "T1_Library1:1958079" "T1_Library1:1958110"
#>  [43] "T1_Library1:1958117" "T1_Library1:1958136" "T1_Library1:1958152"
#>  [46] "T1_Library1:1958169" "T1_Library1:1958213" "T1_Library1:1957974"
#>  [49] "T1_Library1:1958238" "T1_Library1:1958431" "T1_Library1:1958546"
#>  [52] "T1_Library1:1958547" "T1_Library1:1958548" "T1_Library1:1957481"
#>  [55] "T1_Library1:1958235" "T1_Library1:1958550" "T1_Library1:1958554"
#>  [58] "T1_Library1:1958557" "T1_Library1:1958559" "T1_Library1:1957555"
#>  [61] "T1_Library1:1958561" "T1_Library1:1958563" "T1_Library1:1958566"
#>  [64] "T1_Library1:1958568" "T1_Library1:1958206" "T1_Library1:1958268"
#>  [67] "T1_Library1:1958570" "T1_Library1:1958571" "T1_Library1:1958572"
#>  [70] "T1_Library1:1958573" "T1_Library1:1956099" "T1_Library1:1956574"
#>  [73] "T1_Library1:1957047" "T1_Library1:1957061" "T1_Library1:1957240"
#>  [76] "T1_Library1:1957673" "T1_Library1:1957846" "T1_Library1:1957868"
#>  [79] "T1_Library1:1957914" "T1_Library1:1957979" "T1_Library1:1957980"
#>  [82] "T1_Library1:1958008" "T1_Library1:1958058" "T1_Library1:1958195"
#>  [85] "T1_Library1:1958271" "T1_Library1:1958274" "T1_Library1:1958278"
#>  [88] "T1_Library1:1958325" "T1_Library1:1958575" "T1_Library1:1956900"
#>  [91] "T1_Library1:1957651" "T1_Library1:1958245" "T1_Library1:1958386"
#>  [94] "T1_Library1:1958577" "T1_Library1:1958581" "T1_Library1:1958583"
#>  [97] "T1_Library1:1958585" "T1_Library1:1958591" "T1_Library1:1958284"
#> [100] "T1_Library1:1958285" "T1_Library1:1958118" "T1_Library1:1958595"
#> [103] "T1_Library1:1958598" "T1_Library1:1957826" "T1_Library1:1958407"
#> [106] "T1_Library1:1957982" "T1_Library1:1958182" "T1_Library1:1956390"
#> [109] "T1_Library1:1958120" "T1_Library1:1958170" "T1_Library1:1958602"
#> [112] "T1_Library1:1958604" "T1_Library1:1958198" "T1_Library1:1958605"
#> [115] "T1_Library1:1958607" "T1_Library1:1958609" "T1_Library1:1958010"
#> [118] "T1_Library1:1958433" "T1_Library1:1958613" "T1_Library1:1958615"
#> [121] "T1_Library1:1956822" "T1_Library1:1957607" "T1_Library1:1958228"
#> [124] "T1_Library1:1958288" "T1_Library1:1958290" "T1_Library1:1958498"
#> [127] "T1_Library1:1958617" "T1_Library1:1958264" "T1_Library1:1958618"
#> [130] "T1_Library1:1956078" "T1_Library1:1956079" "T1_Library1:1958293"
#> [133] "T1_Library1:1958349" "T1_Library1:1958362" "T1_Library1:1958619"
#> [136] "T1_Library1:1958620" "T1_Library1:1958621" "T1_Library1:1958624"
#> [139] "T1_Library1:1958144" "T1_Library1:1958629" "T1_Library1:1958631"
#> [142] "T1_Library1:1958633" "T1_Library1:1958634" "T1_Library1:1958637"
#> [145] "T1_Library1:1958638" "T1_Library1:1958639" "T1_Library1:1958224"
#> [148] "T1_Library1:1958331" "T1_Library1:1958640" "T1_Library1:1958642"
#> [151] "T1_Library1:1958217" "T1_Library1:1958389" "T1_Library1:1957145"
#> [154] "T1_Library1:1958338" "T1_Library1:1958440" "T1_Library1:1958645"
#> [157] "T1_Library1:1958650" "T1_Library1:1957182" "T1_Library1:1957303"
#> [160] "T1_Library1:1957332" "T1_Library1:1957986" "T1_Library1:1958653"
#> [163] "T1_Library1:1958655" "T1_Library1:1958656" "T1_Library1:1958660"
#> [166] "T1_Library1:1958662" "T1_Library1:1958666" "T1_Library1:1958667"
#> [169] "T1_Library1:1958366" "T1_Library1:1958672" "T1_Library1:1957092"
#> [172] "T1_Library1:1958391" "T1_Library1:1958684" "T1_Library1:1958689"
#> [175] "T1_Library1:1958693" "T1_Library1:1958694" "T1_Library1:1958701"
#> [178] "T1_Library1:1958704" "T1_Library1:1958708" "T1_Library1:1958712"
#> [181] "T1_Library1:1957873" "T1_Library1:1958714" "T1_Library1:1958720"
#> [184] "T1_Library1:1956823" "T1_Library1:1958011" "T1_Library1:1958351"
#> [187] "T1_Library1:1958384" "T1_Library1:1957016" "T1_Library1:1957333"
#> [190] "T1_Library1:1957653" "T1_Library1:1958061" "T1_Library1:1958265"
#> [193] "T1_Library1:1958443" "T1_Library1:1958723" "T1_Library1:1958423"
#> [196] "T1_Library1:1956329" "T1_Library1:1957048" "T1_Library1:1958309"
#> [199] "T1_Library1:1958444" "T1_Library1:1958453" "T1_Library1:1958454"
#> [202] "T1_Library1:1958471" "T1_Library1:1957147" "T1_Library1:1958368"
#> [205] "T1_Library1:1958424" "T1_Library1:1958353" "T1_Library1:1958728"
#> [208] "T1_Library1:1958412" "T1_Library1:1958460" "T1_Library1:1958472"
#> [211] "T1_Library1:1958393" "T1_Library1:1958461" "T1_Library1:1958731"
#> [214] "T1_Library1:1958733" "T1_Library1:1958738" "T1_Library1:1958739"
#> [217] "T1_Library1:1958404" "T1_Library1:1958744" "T1_Library1:1958269"
#> [220] "T1_Library1:1958759" "T1_Library1:1958761" "T1_Library1:1958146"
#> [223] "T1_Library1:1958483" "T1_Library1:1958766" "T1_Library1:1958772"
#> [226] "T1_Library1:1958774" "T1_Library1:1958200" "T1_Library1:1952749"
#> [229] "T1_Library1:1957441" "T1_Library1:1957987" "T1_Library1:1958315"
#> [232] "T1_Library1:1958425" "T1_Library1:1958775" "T1_Library1:1958783"
#> [235] "T1_Library1:1958786" "T1_Library1:1958420" "T1_Library1:1958790"
#> [238] "T1_Library1:1958795" "T1_Library1:1957257" "T1_Library1:1958798"
#> [241] "T1_Library1:1958801" "T1_Library1:1958805" "T1_Library1:1958806"
#> [244] "T1_Library1:1958808" "T1_Library1:1958812" "T1_Library1:1958819"
#> [247] "T1_Library1:1958823" "T1_Library1:1958474" "T1_Library1:1958830"
#> [250] "T1_Library1:1958832" "T1_Library1:1958836" "T1_Library1:1958843"
#> [253] "T1_Library1:1958846" "T1_Library1:1958848" "T1_Library1:1958851"
#> [256] "T1_Library1:1958857" "T1_Library1:1958865" "T1_Library1:1958867"
#> [259] "T1_Library1:1958871" "T1_Library1:1958881" "T1_Library1:1958889"
#> [262] "T1_Library1:1958891" "T1_Library1:1958897" "T1_Library1:1958899"
#> [265] "T1_Library1:1958900" "T1_Library1:1958902" "T1_Library1:1958446"
#> [268] "T1_Library1:1958455" "T1_Library1:1958463" "T1_Library1:1958486"
#> [271] "T1_Library1:1958905" "T1_Library1:1958907" "T1_Library1:1958488"
#> [274] "T1_Library1:1958909" "T1_Library1:1958911" "T1_Library1:1958912"
#> [277] "T1_Library1:1958466" "T1_Library1:1958919" "T1_Library1:1958922"
#> [280] "T1_Library1:1958925" "T1_Library1:1958927" "T1_Library1:1958930"
#> [283] "T1_Library1:1958932" "T1_Library1:1958934" "T1_Library1:1958937"
#> [286] "T1_Library1:1958941" "T1_Library1:1953178" "T1_Library1:1958186"
#> [289] "T1_Library1:1958225" "T1_Library1:1958502" "T1_Library1:1958945"
#> [292] "T1_Library1:1956755" "T1_Library1:1958946" "T1_Library1:1958949"
#> [295] "T1_Library1:1958950" "T1_Library1:1958952" "T1_Library1:1958961"
#> [298] "T1_Library1:1958962" "T1_Library1:1957326" "T1_Library1:1958969"
#> [301] "T1_Library1:1957148" "T1_Library1:1958219" "T1_Library1:1958981"
#> [304] "T1_Library1:1958984" "T1_Library1:1958997" "T1_Library1:1957517"
#> [307] "T1_Library1:1958450" "T1_Library1:1958451" "T1_Library1:1959000"
#> [310] "T1_Library1:1959002" "T1_Library1:1959003" "T1_Library1:1959004"
#> [313] "T1_Library1:1959005" "T1_Library1:1959009" "T1_Library1:1959011"
#> [316] "T1_Library1:1959012" "T1_Library1:1959015" "T1_Library1:1959022"
#> [319] "T1_Library1:1959026" "T1_Library1:1958478" "T1_Library1:1959032"
#> [322] "T1_Library1:1959033" "T1_Library1:1959034" "T1_Library1:1959040"
#> [325] "T1_Library1:1959041" "T1_Library1:1959044" "T1_Library1:1959045"
#> [328] "T1_Library1:1959049" "T1_Library1:1959050" "T1_Library1:1959051"
#> [331] "T1_Library1:1958490" "T1_Library1:1959058" "T1_Library1:1959062"
#> [334] "T1_Library1:1957442"
#> 
#> $`chr14:106327759_A_T`$alt
#> [1] "T1_Library1:1958782"
#> 
#>