Get read counts for a specific base in the genome

Uses samtools pileup to get the read counts for each base in the genomic position specified

get_read_counts(chr, pos, bam, tag = "", min_base_quality = 20,
  max_depth = 1e+05, min_mapq = 30)

Arguments

chr
chromosome name

pos
genomic coordinate

bam
path to bam file

tag
the RG tag if the bam has more than one sample

min_base_quality
minimum base quality for a read to be counted

max_depth
maximum depth above which sampling will happen

min_mapq
the minimum mapping quality for a read to be counted

Value

a list, number of reads for each of the four basepairs